[CHORE] 리뷰반영 - DeadlineManager 객체로 관리 (#59)
